Updated to use tracks.csv for genres
The code has been updated to use the tracks.csv file as a look up to find a song files genre information. Thanks to Joe for this code.
I included a command option to "ReTag" the audio files with the genre discovered in the tracks.csv file as well but this only happens if the genres don't match and "ReTag" option is set. If for some reason the genre isn't in the csv file then the genre from the file meta-data will be used.