Major Update: csv file usage, retagging, image sorting
Did a great deal of clean up and bit of rewriting to make it all flow better. Added commands and functions for just sorting spectrograms or audmages. I also included the csv genre look up (and tested it). All options can be used individually or in combinations. However since the data dir provided at the command line must be for images and not audio the dataset can't be automatically generated after sorting the spectrograms when not creating them. Instead the program will prompt you to restart it with the dataset option invoked. The sort does not happen if you include "dataset" command.