Speed Optimized, added/removed commands
I rewrote this a bit more so that the optimizations I stumbled upon are implemented directly into the main program instead of side scripts. I removed the option of re-tagging the mp3 files for now as it was not really necessary for our project purposes and added another package that needed to be imported. Instead files without matching genre data are simply skipped. I re-arranged some of the functions to better separate and optimize them. The audmage creation code is now fully implemented and enabled! And finally I added two new options TEST and COPY, the test option forces the image creation to stop after just 5 images and move on, instead of 8000. And the copy option tells the audio sorting to copy audio files instead of moving them so you can choose which one is more convenient.