BR2-run Mods and Bug fixes, minor feature update
Fixed some bugs in the program that caused problems on the super computer runs, fixed issue where mp3 file could be read but just didn't contain any data and crashed the program, now skips those files when it finds them. Added code to skip files that already exist. This is very beneficial if you are creating thousands of spectrograms and the program is stopped for whatever reason mid way through.