Keras CNN script w/predictions
This file is the CNN script using Keras. It includes prediction code and saving of predictions as well as command line arguments for changing the directory etc.. The file also includes some graphing and saving of those graphs. The main usage for this file is for training a CNN model on our dataset. Defaults to the medium dataset.